Roulette Game & Vectors
This is a cool little console game that simulates the roulette wheel.  My favorite part is towards the end where I used a multidimensional array to create the board and generate all the sets of winning bids.

The Vectorboy is a little console app that will generate random points in both a 2D plane as well as a 3D space, then provide the shortest distance between the closest two points.